% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{between_cluster_inert}
\alias{between_cluster_inert}
\title{Between-clusters inertia}
\usage{
between_cluster_inert(Z, indices, indices_c, w = rep(1/nrow(Z), nrow(Z)),
  D = rep(1, ncol(Z)))
}
\arguments{
\item{Z}{a numerical data matrix.}

\item{indices}{indices vectors of the first class (\eqn{C_1})}

\item{indices_c}{indices vectors of the second class (\eqn{C_2})}

\item{w}{weights vector}

\item{D}{diagonal distance matrix vector}
}
\description{
Performs the between-clusters inertia of a biparition \eqn{(C_1, C_2)} of a subset of rows of a data matrix.
}
\keyword{internal}

