% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{between_cluster_inert_1D}
\alias{between_cluster_inert_1D}
\title{1D between-clusters inertia}
\usage{
between_cluster_inert_1D(Z, compo, indices, indices_c, w = rep(1/nrow(Z),
  nrow(Z)), D = rep(1, ncol(Z)))
}
\arguments{
\item{Z}{matrix data}

\item{compo}{columns number}

\item{indices}{indices vectors of the first class (\eqn{C_1})}

\item{indices_c}{indices vectors of the second class (\eqn{C_2})}

\item{w}{weights vectors}

\item{D}{diagonal distance matrix vector}
}
\description{
Computes 1D between-clusters inertia of a bipatition \eqn{(C_1, C_2)} of a subset of rows of a data matrix.
}
\keyword{internal}

