% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{compute_cut_values}
\alias{compute_cut_values}
\title{Computes the cut values for a numerical variable}
\usage{
compute_cut_values(col)
}
\arguments{
\item{col}{the column vector in the matrix}
}
\value{
a vector of cut values
}
\description{
Function which computes the cut values of the binary questions defined
with a numerical variables (a column of the data matrix).
}
\seealso{
\link{choose_question}
}
\keyword{internal}

