% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutreediv.R
\name{cutreediv}
\alias{cutreediv}
\title{Cut the tree}
\usage{
cutreediv(tree, K)
}
\arguments{
\item{tree}{the divclust object}

\item{K}{an integer with the desired number of clusters.}
}
\value{
\item{clusters}{the list of observations in each  cluster}

\item{description}{the monothetic description of each cluster}

\item{which_cluster}{a vector of integers indicating the cluster of each observation}

\item{B}{the proportion of inertia explained by the partition (between-cluster inertia/total inertia)}

\item{leaves}{an internal list of \strong{leaves}}
}
\description{
This function cuts the tree into several cluters by specifying the desired number of clusters.
}
\examples{
data(protein) # pure quantitatives data
tree <- divclust(protein) # full clustering
p_5 <- cutreediv(tree,K=5) # partition in 5 clusters
p_5

data(dogs) # pure qualitative data
tree <- divclust(dogs) # full clustering
p_4 <- cutreediv(tree,K=4) # partition in 4 clusters
data(wine) # mixed data
data <- wine[,1:29]
tree <- divclust(data) # full clustering
p_4 <- cutreediv(tree, 4) #
}

