% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.divclust.R
\name{draw}
\alias{draw}
\title{Draw the tree}
\usage{
draw(cluster, x_pos, y_pos, width, height, nqbin, label, threshold, ...)
}
\arguments{
\item{cluster}{object returned by divclust function}

\item{x_pos}{initial abcissa for drawing}

\item{y_pos}{initial ordinate for drawing}

\item{width}{elementary horizontal length for the tree}

\item{height}{elementary vertically length for the tree}

\item{threshold}{cut-off value to decide if the text is written in the tree}

\item{...}{further arguments passed from other methods}
}
\description{
Draw a tree representing the hierarical structure
}
\keyword{internal}

