% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divclust.R
\name{equality_case}
\alias{equality_case}
\title{Equality case data}
\format{A numerial data frame with 20 rows and  2 columns simulated from two gaussien distributions.}
\description{
These data illstrate the case where two binary questions give the same bipartition and how
the binary question with the most discriminant variable (X1 here) is chosen.
}
\examples{
data(equality_case)
plot(equality_case) #X1 discriminates the bipartition better than X2
tree <- divclust(equality_case,K=3)
plot(tree,nqbin=1) # the binary question with X1 is chosen
}

