% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{format_cat_data}
\alias{format_cat_data}
\title{Binary transformation of categorical data}
\usage{
format_cat_data(frame)
}
\arguments{
\item{frame}{a data frame with n rows and p categorical colums (columns are of class factor)}
}
\value{
\item{Y}{an indicator matrix with n rows and m columns where m is the total number of categories}

\item{vec_quali}{a vector of size p containing the number of categories of each variable}

\item{mod_quali}{a list with the set of categories of each variable}
}
\description{
Organizes categorical data into an indicator matrix and gives a description of the categorical variables
}
\keyword{internal}

