% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{gsvd}
\alias{gsvd}
\title{Generalized Singular Value Decomposition}
\usage{
gsvd(A, M, N)
}
\arguments{
\item{A}{a n times p numerical matrix.}

\item{M}{a vector of size p with diag(M)  the metric on \eqn{R^p}.}

\item{N}{a vector of size n with diag(N)  the metric on \eqn{R^n}.}
}
\value{
a 3-tuple \eqn{(U,V,d)} with the left singular vectors, right singular vectors and singular values.
}
\description{
Performs the generalized singular value decomposition \eqn{A=UDV} with weights M on the columns and N on the rows.
}

