% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{inert_1D}
\alias{inert_1D}
\title{1D inertia}
\usage{
inert_1D(Z, compo, indices, w = rep(1/nrow(Z), nrow(Z)), D = rep(1,
  ncol(Z)))
}
\arguments{
\item{Z}{matrix data}

\item{compo}{column number}

\item{indices}{vector representing the subset of rows}

\item{w}{weight vector}

\item{D}{diagonal distance matrix}
}
\description{
Computes the 1D inertia of a column of a matrix data on a subset of rows of a data matrix.
}
\keyword{internal}

