% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{list_leaves}
\alias{list_leaves}
\title{list_leaves}
\usage{
list_leaves(cluster)
}
\arguments{
\item{cluster}{the divclust object}
}
\value{
leaves a list with the list of observations of a cluster ($class) and its monthetic description ($path)

classes a vector of integers indicating the cluster to which each observation is allocated.
}
\description{
This function finds terminal clusters, their monothetic description and the terminal partition.
If the tree is complete, the terminal partition is the partition of singleton (one observation in each cluster)
}
\keyword{internal}

