% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_description}
\alias{make_description}
\title{Description of the leaves}
\usage{
make_description(leaves, cluster)
}
\arguments{
\item{leaves}{a list of leaves}

\item{cluster}{the divclust object containing the tree}
}
\value{
The monothetic description of the leaves
}
\description{
Builds the monothetic description of the leaves
}
\keyword{internal}

