% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_MDI_importance}
\alias{make_MDI_importance}
\title{Computes Mean Decrease of Impurity (MDI Index)}
\usage{
make_MDI_importance(leaves, cluster, weighting = FALSE)
}
\arguments{
\item{leaves}{a list of leaves}

\item{cluster}{the divclust object}

\item{weighting}{Logical (TRUE or FALSE). If TRUE, node inertia is weighted in the calculation of variable importance.}
}
\value{
MDI Importance
}
\description{
Function which computes the MDI Index.
}
