% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_path}
\alias{make_path}
\title{Builds two necessary conditions}
\usage{
make_path(node, path)
}
\arguments{
\item{node}{current node}

\item{path}{of the current node}
}
\value{
the path of the two clusters of the biparition of the current node (left and rigth)
}
\description{
Builds the monothetic description (path) of the two clusters of a bipartition
}
\keyword{internal}

