% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_sentences}
\alias{make_sentences}
\title{Builds two binary questions}
\usage{
make_sentences(cluster, node)
}
\arguments{
\item{cluster}{the divclust object containing the tree}

\item{node}{current node for which we wants to build the questions}
}
\value{
a pair of the two complementaries binary questions (left and rigth)
}
\description{
Builds the two binary questions of a bipartition
}
\keyword{internal}

