% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_utils.R
\name{partitionsequal}
\alias{partitionsequal}
\title{tests if two lists of vectors represent or not the same partition}
\usage{
partitionsequal(l1, l2)
}
\arguments{
\item{l1}{first partition}

\item{l2}{first partition}
}
\value{
a logical
}
\description{
tests if two lists of vectors represent or not the same partition
}
\keyword{internal}

