% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.divclust.R
\name{plot.divclust}
\alias{plot.divclust}
\title{Dendrogram with binary questions}
\usage{
\method{plot}{divclust}(x, nqbin = 3, label = TRUE, ...)
}
\arguments{
\item{x}{object of class divclust}

\item{nqbin}{an integer between 0 and K-1 (K is the number of leaves in the tree). Indicates
the number of binary questions drawn on plot. This parameter is used to plot the binary questions
only of the top levels of the dendrogram. The default value is 3.}

\item{label}{If TRUE, the labels of the observations are drawn.}

\item{...}{further arguments passed from other methods}
}
\description{
Plot the  dendrogram produced by divclust algorithm with the binary questions
associated with each split. The number of binary questions drawn on the dendrogram can be changed
and if the text of the question is to long, is is printed on the console.
}
\examples{
data(protein) # pure quantitatives data
c_tot <- divclust(protein) # full clustering
plot(c_tot)
plot(c_tot,nqbin=4) # the text of the 4th question is printed in the console
c_4 <- divclust(protein, K=4) # stops the clustering to 4 clusters
plot(c_4)

data(dogs) # pure qualitative data
c_tot <- divclust(dogs) # full clustering
plot(c_tot,nqbin=4) # the text of the 4th question is printed in the console

data(wine) # mixed data
data <- wine[,1:29]
c_tot <- divclust(data) # full clustering
plot(c_tot)
}
\seealso{
\link{divclust}
}

