% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.cutreediv.R
\name{print.cutreediv}
\alias{print.cutreediv}
\title{Prints a 'cutdiv' object}
\usage{
\method{print}{cutreediv}(x, ...)
}
\arguments{
\item{x}{object of class cutdiv}

\item{...}{further arguments passed from other methods. They are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class \code{cutdiv}.
}

