% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.divclust.R
\name{print.divclust}
\alias{print.divclust}
\title{Prints a 'divclust' object}
\usage{
\method{print}{divclust}(x, ...)
}
\arguments{
\item{x}{object of class divclust}

\item{...}{further arguments passed from other methods. They are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class \code{divclust}.
}

