% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_mix.R.R
\name{split_mix}
\alias{split_mix}
\title{Splits in quantitative and qualitative datasets}
\usage{
split_mix(base)
}
\arguments{
\item{base}{the data}
}
\value{
\item{data_quanti}{the numerical data frame}

\item{data_quali}{the categorical data frame}
}
\description{
Splits a dataframe in two data frames. The first one contains the numerical columns
and the second one contains the categorical columns.
}
\examples{
data(wine)
split_mix(wine)$data_quanti
split_mix(wine)$data_quali
}

