% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divclust.R
\name{wine}
\alias{wine}
\title{Wines of Val de Loire data}
\format{A data frame with 21 rows (the number of wines) and
31 columns: the first column corresponds to the label of origin,
the second column corresponds to the soil,
and the others correspond to sensory descriptors.}
\source{
Centre de recherche INRA d'Angers
}
\description{
data refering to 21 wines of Val de Loire.
}

